splitted text_ui by separating creation of objects into Create module
